Refactor Preloader: Reorganize imports and extract constants
- Move constants to a separate file
- Update index exports to reflect new file structure
- Remove Preloader.context file and simplify component organization
