+++ src/components/ui/preloader/constants.ts
@@ -1,0 +1,4 @@
+export const END_LOADING_IN = 99;
+export const INITIAL_PERCENT = 1;
+export const LOADING_TIMEOUT = 5000;
+export const EXTERNAL_LOADING_TIMEOUT = LOADING_TIMEOUT + 1000;
